+++ presentation/src/main/java/com/threegap/bitnagil/presentation/onboarding/OnBoardingScreen.kt
@@ -18,8 +18,8 @@ import com.threegap.bitnagil.presentation.onboarding.component.template.OnBoardi
 import com.threegap.bitnagil.presentation.onboarding.model.OnBoardingItem
 import com.threegap.bitnagil.presentation.onboarding.model.OnBoardingPageInfo
 import com.threegap.bitnagil.presentation.onboarding.model.OnBoardingSetType
-import com.threegap.bitnagil.presentation.onboarding.model.mvi.OnBoardingSideEffect
-import com.threegap.bitnagil.presentation.onboarding.model.mvi.OnBoardingState
+import com.threegap.bitnagil.presentation.onboarding.contract.OnBoardingSideEffect
+import com.threegap.bitnagil.presentation.onboarding.contract.OnBoardingState
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
