+++ presentation/src/main/java/com/threegap/bitnagil/presentation/onboarding/contract/OnBoardingSideEffect.kt
@@ -1,0 +1,7 @@
+package com.threegap.bitnagil.presentation.onboarding.contract
+
+sealed interface OnBoardingSideEffect {
+    data object MoveToPreviousScreen : OnBoardingSideEffect
+    data object NavigateToHomeScreen : OnBoardingSideEffect
+    data class ShowToast(val message: String) : OnBoardingSideEffect
+}
