--- presentation/src/main/java/com/threegap/bitnagil/presentation/onboarding/model/mvi/OnBoardingSideEffect.kt
@@ -1,7 +1,0 @@
-package com.threegap.bitnagil.presentation.onboarding.model.mvi
-
-sealed class OnBoardingSideEffect {
-    data object MoveToPreviousScreen : OnBoardingSideEffect()
-    data object NavigateToHomeScreen : OnBoardingSideEffect()
-    data class ShowToast(val message: String) : OnBoardingSideEffect()
-}
